% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtk_cast_matrix.R
\name{mtk_cast_matrix}
\alias{mtk_cast_matrix}
\title{Project data frame as a sparse matrix.}
\usage{
mtk_cast_matrix(dtm, row, column, value, ...)
}
\arguments{
\item{dtm}{}
}
\value{
A sparse matrix
}
\description{
Project data frame as a sparse matrix.
}
