% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtk_dtm_tcm.R
\name{mtk_dtm_tcm}
\alias{mtk_dtm_tcm}
\title{Convet DTM to TCM; ie, treat document as context.}
\usage{
mtk_dtm_tcm(dtm)
}
\arguments{
\item{dtm}{A document-term matrix}
}
\value{
A sparse matrix
}
\description{
Convet DTM to TCM; ie, treat document as context.
}
