% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtk_summarize_lda.R
\name{mtk_summarize_lda}
\alias{mtk_summarize_lda}
\title{Quick:
Summary:}
\usage{
mtk_summarize_lda(lda, topic_feats_n = 10)
}
\arguments{
\item{per}{parameter}
}
\value{

}
\description{
Quick:
Summary:
}
