% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_2d.R
\name{pmtk_2d}
\alias{pmtk_2d}
\title{Build data structures 2d visualizations.}
\usage{
pmtk_2d(mat, d.method = "euclidean", c.method = "ward.D")
}
\arguments{
\item{x}{A topic-term summary df.}
}
\value{
A list.
}
\description{
Build data structures 2d visualizations.
}
