% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_batch_abstracts.R
\name{pmtk_batch_abstracts}
\alias{pmtk_batch_abstracts}
\title{Download abstract and meta data for research articles included in PubMed.}
\usage{
pmtk_batch_abstracts(pmids, out_file, file_prefix, keep_empty_abs = FALSE)
}
\arguments{
\item{pmids}{A vector of PMIDs}

\item{out_file}{File path for output}

\item{file_prefix}{String specifying batch name}

\item{keep_empty_abs}{Boolean; FALSE excludes records for which abstract is not included}
}
\value{
A set of data frames as .RDS files
}
\description{
Download abstract and meta data for research articles included in PubMed.
}
