% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_build_interactive.R
\name{pmtk_build_interactive}
\alias{pmtk_build_interactive}
\title{Build interactive html for topic summary}
\usage{
pmtk_build_interactive(pmtk_lda, pmtk_2d, out_dir, file_name)
}
\arguments{
\item{pmtk_lda}{Object returned from pmtk_summarize_lds}

\item{pmtk_2d}{Object returned from pmtk_2d}

\item{out_dir}{Directory to output html}

\item{file_name}{Name to assign html file}
}
\value{
An html file
}
\description{
Build interactive html for topic summary
}
