% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_cast_spmatrix.R
\name{pmtk_cast_spmatrix}
\alias{pmtk_cast_spmatrix}
\title{Project data frame as a sparse matrix.}
\usage{
pmtk_cast_spmatrix(dtm, row, column, value, ...)
}
\arguments{
\item{dtm}{}
}
\value{
A sparse matrix
}
\description{
Project data frame as a sparse matrix.
}
