% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_crosstab_query.R
\name{pmtk_crosstab_query}
\alias{pmtk_crosstab_query}
\title{Download abstract and meta data for research articles included in PubMed.}
\usage{
pmtk_crosstab_query(search_results)
}
\arguments{
\item{search_results}{A df returned from pmtk_search_pubmed()}
}
\value{
A data frame
}
\description{
Download abstract and meta data for research articles included in PubMed.
}
