% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_download_abs.R
\name{pmtk_download_abs}
\alias{pmtk_download_abs}
\title{Download abstract and meta data for research articles included in PubMed.}
\usage{
pmtk_download_abs(pmids, out_file, file_prefix)
}
\arguments{
\item{pmids}{A vector of PMIDs}

\item{out_file}{File path for output}

\item{file_prefix}{String specifying batch name}
}
\value{
A set of data frames as .RDS files
}
\description{
Download abstract and meta data for research articles included in PubMed.
}
