% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_dtm_tcm.R
\name{pmtk_dtm_tcm}
\alias{pmtk_dtm_tcm}
\alias{pmtk_dtm2tcm}
\title{Convet DTM to TCM; ie, treat document as context.}
\usage{
pmtk_dtm2tcm(dtm)
}
\arguments{
\item{dtm}{}
}
\value{
A sparse matrix
}
\description{
Convet DTM to TCM; ie, treat document as context.
}
