% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_extract_pmc.R
\name{pmtk_extract_pmc}
\alias{pmtk_extract_pmc}
\title{Structure PMC article as data frame, grouped by article section}
\usage{
pmtk_extract_pmc(x)
}
\arguments{
\item{x}{A file path to a PMC article}
}
\value{
A data frame
}
\description{
Structure PMC article as data frame, grouped by article section
}
