% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_gather_mesh.R
\name{pmtk_gather_mesh}
\alias{pmtk_gather_mesh}
\title{Extract MeSH terms.}
\usage{
pmtk_gather_mesh(x)
}
\arguments{
\item{x}{A data frame returned from pmtk_get_records2()}
}
\value{
A data frame
}
\description{
Extract MeSH terms.
}
