% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_get_ids.R
\name{pmtk_get_pmids}
\alias{pmtk_get_pmids}
\title{Perform basic PubMed search.}
\usage{
pmtk_get_pmids(pmed_search, convert_syntax = T, verbose = T)
}
\arguments{
\item{pmids}{A vector of PMIDs}

\item{out_file}{File path for output}

\item{file_prefix}{String specifying how batch file name}
}
\value{
A data frame of PMIDs
}
\description{
Perform basic PubMed search.
}
