% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_get_records.R
\name{pmtk_get_records}
\alias{pmtk_get_records}
\title{Download abstract and meta data for research articles included in PubMed.}
\usage{
pmtk_get_records(pmids, cores)
}
\arguments{
\item{pmids}{A vector of PMIDs}

\item{cores}{Numeric specifying number of cores to use}
}
\value{
A list of data frames
}
\description{
Download abstract and meta data for research articles included in PubMed.
}
