% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_get_records1.R
\name{pmtk_get_records1}
\alias{pmtk_get_records1}
\title{Download abstract and meta data for research articles included in PubMed.}
\usage{
pmtk_get_records1(x)
}
\arguments{
\item{x}{A character vector of PMIDs}
}
\value{
A data frame.
}
\description{
Download abstract and meta data for research articles included in PubMed.
}
