% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_get_records2.R
\name{pmtk_get_records2}
\alias{pmtk_get_records2}
\title{Download abstract and meta data from PubMed.}
\usage{
pmtk_get_records2(pmids, cores = 3, ncbi_key = NULL)
}
\arguments{
\item{pmids}{A vector of PMIDs}

\item{cores}{Numeric specifying number of cores to use}

\item{ncbi_key}{API key}
}
\value{
A list of data frames
}
\description{
Download abstract and meta data from PubMed.
}
