% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_loadr_abs.R
\name{pmtk_loadr_abs}
\alias{pmtk_loadr_abs}
\title{Load output from `pmtk_download_abs` as a single data.frame}
\usage{
pmtk_loadr_abs(in_file, file_prefix)
}
\arguments{
\item{in_file}{File path to batches}

\item{file_prefix}{String specifying batch name}
}
\value{
A data.frame
}
\description{
Load output from `pmtk_download_abs` as a single data.frame
}
