% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_locate_search.R
\name{pmtk_locate_search}
\alias{pmtk_locate_search}
\title{Search corpus for lexico-gram patterns in context via corpus::text_locate function.
Details.}
\usage{
pmtk_locate_search(text, doc_id, search, stem = F, window = 15)
}
\arguments{
\item{text}{character vector}

\item{doc_id}{character vector}

\item{search}{string}

\item{stem}{boolean}

\item{window}{integer}
}
\value{
A data frame.
}
\description{
Search corpus for lexico-gram patterns in context via corpus::text_locate function.
Details.
}
