% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_locate_term.R
\name{pmtk_locate_term}
\alias{pmtk_locate_term}
\title{Search corpus for lexical patterns in context via corpus::text_locate function.}
\usage{
pmtk_locate_term(text, doc_id, term, stem = F, window = 15)
}
\arguments{
\item{text}{character vector}

\item{doc_id}{character vector}

\item{term}{string}

\item{stem}{boolean}

\item{window}{integer}
}
\value{
A data frame.
}
\description{
Search corpus for lexical patterns in context via corpus::text_locate function.
}
