% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_query_bigrams.R
\name{pmtk_query_bigrams}
\alias{pmtk_query_bigrams}
\title{Aggregate search results by query term combinations.}
\usage{
pmtk_query_bigrams(search_results)
}
\value{
A data frame
}
\description{
Aggregate search results by query term combinations.
}
