% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_search_pubmed.R
\name{pmtk_search_pubmed}
\alias{pmtk_search_pubmed}
\title{Perform basic PubMed search.}
\usage{
pmtk_search_pubmed(
  pmed_search,
  translate_syntax = T,
  verbose = T,
  retmax = 5e+06
)
}
\arguments{
\item{pmed_search}{A character string, or list of character strings}

\item{translate_syntax}{boolean: T to translate term to NCBI/mesh+tiab}

\item{verbose}{boolean: T to output progress}
}
\value{
A data frame of PMIDs
}
\description{
Perform basic PubMed search.
}
