% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_strip_batches.R
\name{pmtk_strip_batches}
\alias{pmtk_strip_batches}
\alias{pubmed_strip_batches}
\title{Load output from `pmtk_batch_abstracts` as a single data.frame}
\usage{
pubmed_strip_batches(in_file, file_prefix)
}
\arguments{
\item{in_file}{File path to batches}

\item{file_prefix}{String specifying batch name}
}
\value{
A data.frame
}
\description{
Load output from `pmtk_batch_abstracts` as a single data.frame
}
