% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_strip_xml.R
\name{pmtk_strip_xml}
\alias{pmtk_strip_xml}
\title{Extract text from xml nodes.}
\usage{
pmtk_strip_xml(x)
}
\arguments{
\item{x}{A character vector of PMIDs}
}
\value{
A data frame
}
\description{
Extract text from xml nodes.
}
