% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_strip_xml1.R
\name{pmtk_strip_xml1}
\alias{pmtk_strip_xml1}
\title{Download abstract and meta data for research articles included in PubMed.}
\usage{
pmtk_strip_xml1(x)
}
\arguments{
\item{x}{XML output from NCBI API query.}
}
\value{
A list of data frames
}
\description{
Download abstract and meta data for research articles included in PubMed.
}
