% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmtk_tbl_pmc_ref}
\alias{pmtk_tbl_pmc_ref}
\title{MeSH annotation frequencies.}
\format{
A data frame with 1304697 rows and 4 variables:
}
\source{
\url{https://github.com/jaytimm/PubmedMTK/blob/main/mds/pmc-reference.md}
}
\usage{
data(pmtk_tbl_pmc_ref)
}
\description{
A dataset containing the frequency/relative frequency of MeSH annotations in PubMed abstracts included in the PubMedCentral - Open Access Common Use PMID list.
}
\keyword{datasets}
