% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_toke_sentences.R
\name{pmtk_toke_sentences}
\alias{pmtk_toke_sentences}
\title{A sentence tokenizer that mostly works.}
\usage{
pmtk_toke_sentences(text, doc_id)
}
\arguments{
\item{text}{character vector}

\item{doc_id}{character vector}
}
\value{
A data frame.
}
\description{
A sentence tokenizer that mostly works.
}
