% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtk_summarize_lda.R
\name{mtk_summarize_lda}
\alias{mtk_summarize_lda}
\title{Summarize topic model output
LDA model returned from `text2vec::LDA` - a fast/efficient LDA implementation - excellent for exploration -}
\usage{
mtk_summarize_lda(lda, topic_feats_n = 10)
}
\arguments{
\item{lda}{An lda object returned from `text2vec::LDA`}

\item{topic_feats_n}{Number of features to include in topic summary}
}
\value{

}
\description{
Summarize topic model output
LDA model returned from `text2vec::LDA` - a fast/efficient LDA implementation - excellent for exploration -
}
