% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_crosstab_query.R
\name{pmtk_crosstab_query}
\alias{pmtk_crosstab_query}
\title{Cross-tab search results for multiple calls to pmtk_search_pubmed().}
\usage{
pmtk_crosstab_query(x)
}
\arguments{
\item{x}{A df or list of dfs returned from pmtk_search_pubmed()}
}
\value{
A data frame
}
\description{
Cross-tab search results for multiple calls to pmtk_search_pubmed().
}
