% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_crosstab_query.R
\name{pmtk_crosstab_query}
\alias{pmtk_crosstab_query}
\title{Aggregate search results by query term combinations.}
\usage{
pmtk_crosstab_query(search_results)
}
\value{
A data frame
}
\description{
Aggregate search results by query term combinations.
}
