% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_gather_mesh.R
\name{pmtk_gather_mesh}
\alias{pmtk_gather_mesh}
\title{Extract MeSH terms.}
\usage{
pmtk_gather_mesh(x)
}
\arguments{
\item{meta_df}{Metadata data frame returned from `pmtk_loadr_abs'`}
}
\value{
A data frame
}
\description{
Extract MeSH terms.
}
