% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_gather_mesh.R
\name{pmtk_gather_mesh}
\alias{pmtk_gather_mesh}
\title{Download abstract and meta data for research articles included in PubMed.}
\usage{
pmtk_gather_mesh(meta_df)
}
\arguments{
\item{meta_df}{A vector of PMIDs}
}
\value{
A data frame
}
\description{
Download abstract and meta data for research articles included in PubMed.
}
