% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_search_pubmed.R
\name{pmtk_search_pubmed}
\alias{pmtk_search_pubmed}
\title{Perform basic PubMed search.}
\usage{
pmtk_search_pubmed(pmed_search, verbose = T)
}
\arguments{
\item{pmids}{A vector of PMIDs}

\item{out_file}{File path for output}

\item{file_prefix}{String specifying how batch file name}
}
\value{
A data frame of PMIDs
}
\description{
Perform basic PubMed search.
}
