% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtk_search_pubmed.R
\name{pmtk_search_pubmed}
\alias{pmtk_search_pubmed}
\title{Perform basic PubMed search.}
\usage{
pmtk_search_pubmed(search_term, fields = c("TIAB", "MH"), sleep = 1)
}
\arguments{
\item{search_term}{Query term as character string}

\item{fields}{PubMed fields to query}

\item{sleep}{In seconds.}
}
\value{
A data frame of PMIDs
}
\description{
Perform basic PubMed search.
}
